integrate sk chat agent with app
